requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
Django==4.2
django-admin-sortable2==2.2.1
django-nested-admin==4.0.2
django-reversion==5.0.12
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.7.4
pydantic_core==2.18.4
PyMySQL==1.1.1
python-dotenv==1.0.1
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
yarl==1.9.4
